+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add new `Capybara/VisibilityMatcher` cop. ([@aried3r][])
 * Ignore String constants by `RSpec/Describe`. ([@AlexWayfer][])
 * Drop support for ruby 2.3. ([@bquorning][])
+* Fix multiple cops to detect `let` with proc argument. ([@tejasbubane][])
 
 ## 1.38.1 (2020-02-15)
 
@@ -497,3 +498,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@eitoball]: https://github.com/eitoball
 [@aried3r]: https://github.com/aried3r
 [@AlexWayfer]: https://github.com/AlexWayfer
+[@tejasbubane]: https://github.com/tejasbubane